requirements.txt
pyjpt>=0.1.25
igraph
numpy
plotly
dash_bootstrap_components
dash_daq
watchdog
